### 2023-09-20 #5856 @kyleniemeyer
Hello @Muhao-Chen, it looks like you indicated that the JOSS paper was on the `JOSS_paper` branch, but there is no such branch in the repository. The paper looks to be on the `main` branch, just in the `JOSS_paper` directory.

Also, while we are getting started, can you please explain in detail the differences between this submission and the previous submission ["TsgFEM: Tensegrity Finite Element Method"](https://joss.theoj.org/papers/10.21105/joss.03390)? I noticed that some of the files in the repository for the current position seem copied from TsgFEM.
