+++ JOSS_paper/paper.bib
@@ -6,8 +6,9 @@
   number={13},
   pages={2000636},
   year={2021},
-  publisher={Wiley Online Library}
-
+  publisher={Wiley Online Library},
+doi={https://doi.org/10.1002/advs.202000636}
+}
 @article{chen2019autonomous,
   title={Autonomous deployment of a solar panel using elastic origami and distributed shape-memory-polymer actuators},
   author={Chen, Tian and Bilal, Osama R and Lang, Robert and Daraio, Chiara and Shea, Kristina},
@@ -16,9 +17,9 @@
   number={6},
   pages={064069},
   year={2019},
-  publisher={APS}
+  publisher={APS},
+doi={https://doi.org/10.1103/PhysRevApplied.11.064069}
 }
-
 @article{li2019origami,
   title={Origami concave tubes for energy absorption},
   author={Li, Yang and You, Zhong},
@@ -26,9 +27,9 @@
   volume={169},
   pages={21--40},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.ijsolstr.2019.03.026}
 }
-
 @article{ze2022soft,
   title={Soft robotic origami crawler},
   author={Ze, Qiji and Wu, Shuai and Nishikawa, Jun and Dai, Jize and Sun, Yue and Leanza, Sophie and Zemelka, Cole and Novelino, Larissa S and Paulino, Glaucio H and Zhao, Ruike Renee},
@@ -37,24 +38,23 @@
   number={13},
   pages={eabm7834},
   year={2022},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+doi={https://doi.org/10.1126/sciadv.abm7834}
 }
-
 @article{hernandez2019active,
   title={Active origami},
   author={Hernandez, EA Peraza and Hartl, Darren J and Lagoudas, Dimitris C},
   journal={Active Origami},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+doi={https://doi.org/10.1007/978-3-319-91866-2}
 }
-
 @book{lang2012origami,
   title={Origami design secrets: mathematical methods for an ancient art},
   author={Lang, Robert J},
   year={2012},
   publisher={CRC Press}
 }
-
 @article{huang2020complex,
   title={Complex SiOC ceramics from 2D structures by 3D printing and origami},
   author={Huang, Kai and Elsayed, Hamada and Franchin, Giorgia and Colombo, Paolo},
@@ -62,9 +62,9 @@
   volume={33},
   pages={101144},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.addma.2020.101144}
 }
-
 @article{zhang2019fold,
   title={Fold 2D woven DNA origami to origami+ structures},
   author={Zhang, Yiyang and Wang, Chao and Dong, Yuanchen and Wang, Dianming and Cao, Tianyang and Wang, Shuo and Liu, Dongsheng},
@@ -73,9 +73,9 @@
   number={22},
   pages={1809097},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+doi={https://doi.org/10.1002/adfm.201809097}
 }
-
 @article{ge2014active,
   title={Active origami by 4D printing},
   author={Ge, Qi and Dunn, Conner K and Qi, H Jerry and Dunn, Martin L},
@@ -84,9 +84,9 @@
   number={9},
   pages={094007},
   year={2014},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+doi={10.1088/0964-1726/23/9/094007}
 }
-
 @article{zhao20183d,
[…]
