+++ JOSS_paper/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 
 To fully grasp the foundational concept of integrating origami and tensegrity frameworks, it is imperative to first understand the intrinsic characteristics, advantages, and limitations of each individual paradigm. Originating from the ancient traditions of Chinese paper invention and Japanese artistry, the art form known as origami combines the Japanese terms "oru" (to fold) and "kami" (paper) [@meloni2021engineering]. Initially celebrated in Asian ceremonies like weddings and festivals, origami gained global recognition in the 1950s through renowned artist Akira Yoshizawa. Yoshizawa authored 18 books and crafted more than 50,000 models, dramatically elevating the art form's status [@meloni2021engineering]. Over the years, rigorous academic research has transitioned origami from a purely aesthetic discipline to an engineering asset used in applications like space solar panels [@chen2019autonomous], energy absorption systems [@li2019origami], and robotics [@ze2022soft]. The advantages of origami systems are manifold: (1). Complexity of Form: Origami allows the creation of intricate shapes through complex folding techniques [@hernandez2019active;@lang2012origami]. (2). Compactability: Two-dimensional sheets can be easily compressed into compact forms and later expanded into voluminous structures [@huang2020complex;@zhang2019fold]. (3). Cost-Efficient Manufacturing: The transformation of 2D sheets into 3D or 4D structures simplifies the assembly process, making it highly cost-effective [@ge2014active;@zhao20183d]. (4). Metamaterial Characteristics: Many origami forms feature tunable properties like negative Poisson's ratios, structural bistability, and self-locking capabilities [@kamrava2017origami;@yasuda2015reentrant]. However, the system is not without its limitations: (1). Low Stiffness: While actuators or locking mechanisms can be added to increase stiffness, these modifications necessitate additional design complexity. (2). Control Challenges: The transient dynamics during shape morphing are difficult to manage. While actuation at the hinges can partially mitigate this, it adds further intricacy to the design.
 
-The concept of tensegrity was initially formulated by artists Ioganson in 1921 and Snelson in 1948, showcasing a structural framework comprising compressible elements like bars or struts and tensile components such as strings or cables [@snelson1965continuous]. The term "tensegrity" itself was coined by Buckminster Fuller in 1959, melding 'tensile' and 'integrity' [@fuller1959us]. Subsequent decades of research have demonstrated tensegrity's immense utility in crafting lightweight structures, deployable platforms, and robotics, with applications spanning space platforms [@roffman2021morphing], cable domes [@ma2020design], and space landers [@kim2020rolling]. Strengths of Tensegrity Systems include: (1). Mass Efficiency: Due to axial loading on all one-dimensional structural elements, the system offers lightweight design solutions [@chen2023minimal;@fraddosio2019minimal]. (2). Model Accuracy: The absence of bar bending allows for highly accurate structural models, enabling precise control [@kan2021novel;@ma2022dynamics]. (3). Adjustable Parameters: Structure parameters such as prestress levels and string lengths can be adjusted to achieve varying equilibrium states, stiffness, and shapes [@feng2020optimal;@trinh2022force]. (4). Metamaterial Properties: The configuration of bar-string patterns can yield materials with unique properties [@amendola2019optimal;@bauer2021tensegrity]. (5). Morphological Control: Actuation of the strings allows for substantial shape morphing [@rhodes2019compact;@zhou2021distributed]. Limitations of Tensegrity Systems: (1). Lack of Enclosure: The structure is typically open and can be augmented with membranes or shells, though research in this domain remains limited. (2). Complex Joints: Given that a single node may connect multiple bars, each with varying degrees of freedom, the creation of joints tends to involve non-standard types.
[…]
