+++ JOSS_paper/paper.bib
@@ -326,7 +326,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={205},
   pages={114119},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.bios.2022.114119}
 }
 @article{kresling2012origami,
   title={Origami-structures in nature: lessons in designing “smart” materials},
@@ -335,7 +336,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={1420},
   pages={mrsf11--1420},
   year={2012},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+doi={https://doi.org/10.1557/opl.2012.536}
 }
 @article{faber2018bioinspired,
   title={Bioinspired spring origami},
@@ -345,7 +347,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={6382},
   pages={1386--1391},
   year={2018},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+doi={https://doi.org/10.1126/science.aap7753}
 }
 @inproceedings{rohmer2015experimental,
   title={An experimental and numerical study of shape memory alloy-based tensegrity/origami structures},
@@ -354,7 +357,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={57526},
   pages={V009T12A064},
   year={2015},
-  organization={American Society of Mechanical Engineers}
+  organization={American Society of Mechanical Engineers},
+doi={https://doi.org/10.1115/IMECE2015-51928}
 }
 @article{miranda2020mechanics,
   title={Mechanics of smart origami sunscreens with energy harvesting ability},
@@ -363,7 +367,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={105},
   pages={103503},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.mechrescom.2020.103503}
 }
 @incollection{hong2021tensegami,
   title={Tensegami: Design principle of combining tensegrity and origami to make geodesic dome structure for martian agriculture},
