+++ JOSS_paper/paper.bib
@@ -317,7 +317,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={8},
   pages={1510},
   year={2020},
-  publisher={MDPI}
+  publisher={MDPI},
+doi={https://doi.org/10.3390/nano10081510}
 }
 @article{caratelli2022paper,
   title={A paper-based electrochemical device for the detection of pesticides in aerosol phase inspired by nature: A flower-like origami biosensor for precision agriculture},
