+++ JOSS_paper/paper.bib
@@ -214,7 +214,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={94},
   pages={104584},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.euromechsol.2022.104584}
 }
 @incollection{amendola2019optimal,
   title={Optimal prestress design of the band gap dynamics in tensegrity metamaterials},
@@ -232,7 +233,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={10},
   pages={2005647},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+doi={https://doi.org/10.1002/adma.202005647}
 }
 @article{rhodes2019compact,
   title={Compact shape morphing tensegrity robots capable of locomotion},
@@ -241,7 +243,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={6},
   pages={111},
   year={2019},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+doi={https://doi.org/10.3389/frobt.2019.00111}
 }
 @article{zhou2021distributed,
   title={Distributed actuation and control of a tensegrity-based morphing wing},
@@ -251,7 +254,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={1},
   pages={34--45},
   year={2021},
-  publisher={IEEE}
+  publisher={IEEE},
+doi={https://doi.org/10.1109/TMECH.2021.3058074}
 }
 @article{liu2017nonlinear,
   title={Nonlinear mechanics of non-rigid origami: an efficient computational approach},
@@ -261,7 +265,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={2206},
   pages={20170348},
   year={2017},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+doi={https://doi.org/10.1098/rspa.2017.0348}
 }
 @article{filipov2017bar,
   title={Bar and hinge models for scalable analysis of origami},
@@ -270,7 +275,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={124},
   pages={26--45},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.ijsolstr.2017.05.028}
 }
 @article{chen2020general,
   title={A general approach to minimal mass tensegrity},
@@ -278,7 +284,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   journal={Composite Structures},
   pages={112454},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.compstruct.2020.112454}
 }
 @article{ma2022tensegrity,
   title={Tensegrity system dynamics based on finite element method},
@@ -287,7 +294,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   volume={280},
   pages={114838},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+doi={https://doi.org/10.1016/j.compstruct.2021.114838}
 }
 @article{wang2001mechanical,
   title={Mechanical behavior in living cells consistent with the tensegrity model},
@@ -297,7 +305,8 @@ doi={https://doi.org/10.1016/j.engstruct.2022.114391}
   number={14},
   pages={7765--7770},
   year={2001},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
[…]
