+++ JOSS_paper/paper.md
@@ -1,60 +1,47 @@
 ---
-
 title: 'OriTsgEFA: Origami and Tensegrity Equilibrium and Form-finding Analysis'
-
-
 tags:
   - Origami and Tensegrity systems
   - Flexible structures
   - Prestressable structures
   - Elastic and plastic deformation
   - Stiffness of origami and tensegrity structures 
-
 authors:
   - name: Shuo Ma
     orcid: 0000-0003-3789-2893
     affiliation: 1
-
   - name: Muhao Chen^[corresponding author]
     orcid: 0000-0003-1812-6835
     affiliation: 2
-
   - name: Robert E. Skelton
     orcid: 0000-0001-6503-9115
     affiliation: 3
-    
 affiliations:
  - name: College of Civil Engineering, Zhejiang University of Technology, Hangzhou, Zhejiang, China
    index: 1
-
  - name: Department of Mechanical and Aerospace Engineering, University of Kentucky, Lexington, KY, USA
    index: 2
-
  - name: Department of Aerospace Engineering, Texas A&M University, College Station, TX, USA
    index: 3
-
 date: 31 August 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-To fully grasp the foundational concept of integrating origami and tensegrity frameworks, it is imperative to first understand the intrinsic characteristics, advantages, and limitations of each individual paradigm. Originating from the ancient traditions of Chinese paper invention and Japanese artistry, the art form known as origami combines the Japanese terms "oru" (to fold) and "kami" (paper) [@meloni2021engineering]. Initially celebrated in Asian ceremonies like weddings and festivals, origami gained global recognition in the 1950s through renowned artist Akira Yoshizawa. Yoshizawa authored 18 books and crafted more than 50,000 models, dramatically elevating the art form's status [@meloni2021engineering]. Over the years, rigorous academic research has transitioned origami from a purely aesthetic discipline to an engineering asset used in applications like space solar panels [@chen2019autonomous], energy absorption systems [@li2019origami], and robotics [@ze2022soft]. The advantages of origami systems are manifold: (1). Complexity of Form: Origami allows the creation of intricate shapes through complex folding techniques [@hernandez2019active;@lang2012origami]. (2). Compactability: Two-dimensional sheets can be easily compressed into compact forms and later expanded into voluminous structures [@huang2020complex;@zhang2019fold]. (3). Cost-Efficient Manufacturing: The transformation of 2D sheets into 3D or 4D structures simplifies the assembly process, making it highly cost-effective [@ge2014active;@zhao20183d]. (4). Metamaterial Characteristics: Many origami forms feature tunable properties like negative Poisson's ratios, structural bistability, and self-locking capabilities [@kamrava2017origami;@yasuda2015reentrant]. However, the system is not without its limitations: (1). Low Stiffness: While actuators or locking mechanisms can be added to increase stiffness, these modifications necessitate additional design complexity. (2). Control Challenges: The transient dynamics during shape morphing are difficult to manage. While actuation at the hinges can partially mitigate this, it adds further intricacy to the design.
[…]
