### 2023-09-20 #5856 @Muhao-Chen
> Hello @Muhao-Chen, it looks like you indicated that the JOSS paper was on the `JOSS_paper` branch, but there is no such branch in the repository. The paper looks to be on the `main` branch, just in the `JOSS_paper` directory.
> Thanks for helping us. I have fixed the problem. The pdf looks good now.
>
> Also, while we are getting started, can you please explain in detail the differences between this submission and the previous submission ["TsgFEM: Tensegrity Finite Element Method"](https://joss.theoj.org/papers/10.21105/joss.03390)? I noticed that some of the files in the repository for the current position seem copied from TsgFEM.
> Yes, the TsgFEM facilitates static and dynamic analyses via the finite element method (FEM) for tensegrity systems simulations. On the other hand, OriTsgFEA provides a comprehensive platform for both equilibrium and form-finding studies of combined origami and tensegrity systems. Some of the .m files (i.e., generate the connectivity matrices, the material definition of tensegrity bars and strings) are copied from TsgFEM; these source codes are appropriately cited.
