### 2023-09-22 #5856 @Muhao-Chen
@kyleniemeyer Thanks for your question. There are about 12/35 (TsgFEM/OriTsgEFA). Both are independent software packages for different analysis purposes. The shared files are more like generating pictures, videos, material libraries, etc. Many thanks!
