### 2023-10-11 #5939 @editorialbot
[REVIEW]: OriTsgEFA: Origami and Tensegrity Equilibrium and Form-finding Analysis
**Submitting author:** @Muhao-Chen (<a href="http://orcid.org/0000-0003-1812-6835">Muhao Chen</a>)
**Repository:** https://github.com/Muhao-Chen/Origami_and_Tensegrity_Equilibrium_and_Form-finding_Analysis_OriTsgEFA
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1
