### 2023-10-19 #5939 @Kevin-Mattheus-Moerman
@Muhao-Chen I have removed myself as reviewer here, and am talking to you now instead in my capacity as one of JOSS' Editors in Chief.

Thanks for this submission, and this related submission as well ([TsgFEM](https://doi.org/10.21105/joss.03390)). I have in the past edited or reviewed some of your work, and have found the tensegrity work interesting and useful. However, at this point your team have previously published the following two tensegrity related papers in JOSS: [MOTES](https://doi.org/10.21105/joss.03390), and [TsgFEM](https://doi.org/10.21105/joss.03390). In addition, your direct co-workers have published the closely related paper on [STEDY](https://doi.org/10.21105/joss.01042).

Currently you are seeking the publication of this work here, OriTsgEFA, as well as a project named [TsgDMC](https://github.com/openjournals/joss-reviews/issues/5908).

Having been involved in the review/editing of the previously published works I have just noticed a concern, which I have discussed with the editorial board at length as well. The concern is that although these software projects are each perhaps functionally interesting, they do share a significant amount of functional overlap.

Below are some examples of particular codes. For instance`tenseg_plot.m`, which is common to 4 of your repositories/projects:

1. https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/tenseg_plot.m
2. https://github.com/ramaniitrgoyal92/Modeling_of_Tensegrity_Structures_MOTES/blob/master/Function_Library/tenseg_plot.m
3. https://github.com/Muhao-Chen/Origami_and_Tensegrity_Equilibrium_and_Form-finding_Analysis_OriTsgEFA/blob/main/Function_library/tenseg_plot.m
4. https://github.com/Muhao-Chen/Tensegrity_Data-based_Modeling_and_Control_TsgDMC/blob/main/Software_Verification_and_Examples/SysCtrl_Examples/02_tenseg_airfoil/tenseg_airfoil_dynamics/tenseg_plot.m

Also this `stress_strain.m` function looks very similar:
1. https://github.com/Muhao-Chen/Tensegrity_Data-based_Modeling_and_Control_TsgDMC/blob/main/Software_Verification_and_Examples/SysCtrl_Examples/02_tenseg_airfoil/tenseg_airfoil_dynamics/stress_strain.m
2. https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/stress_strain.m
3. https://github.com/Muhao-Chen/Origami_and_Tensegrity_Equilibrium_and_Form-finding_Analysis_OriTsgEFA/blob/main/Function_library/stress_strain.m

And these solver codes look like they were based on each other:
1. https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/static_solver2.m
2. https://github.com/Muhao-Chen/Origami_and_Tensegrity_Equilibrium_and_Form-finding_Analysis_OriTsgEFA/blob/main/Function_library/static_solver_ori_2.m

Similarly for these `material_info` codes:
1. https://github.com/Muhao-Chen/Tensegrity_Data-based_Modeling_and_Control_TsgDMC/blob/main/Software_Verification_and_Examples/SysCtrl_Examples/02_tenseg_airfoil/tenseg_airfoil_dynamics/material_info2.m
2. https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/material_info.m

And these `ode4...` files:
1. https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/ode4_tsg.m
2. https://github.com/Muhao-Chen/Tensegrity_Data-based_Modeling_and_Control_TsgDMC/blob/main/Software_Verification_and_Examples/SysCtrl_Examples/02_tenseg_airfoil/tenseg_airfoil_dynamics/ode4_truss.m

There are probably more similarities and shared functionalities/overlap if we study these a bit deeper.

Note that starting off from a previous project can happen in practice. However, the correct approach is to fork a project, such that the origin and original authorships for each contribution remain preserved. That way the incremental achievements on top of that fork are also clear. I think in your case the projects seem to be copy-pasted on a local machine. Or that each time you start a project you copy bits-and-pieces from the others as a starting point. Next the "new project" is uploaded to GitHub each time as if a complete new achievement. Here for instance: https://github.com/Muhao-Chen/Tensegrity_Data-based_Modeling_and_Control_TsgDMC/commits?author=Muhao-Chen, the 5 commits show the copied upload (albeit with some new functional developments for which I understand you seek publication) allong with commits to add/update/fix the paper and readme. This is not the proper way to do software versioning/branching/forking. Note I do realise you have cited the previous works in each subsequent publication, however on the Git/GitHub side all usual breadcrumbs to show novelty/changes/authorship has, due to the above approach, been lost.

JOSS does not accept so-called "salami-slicing", i.e. where a potentially large project is avoided, and instead a splintered set of several similar smaller ones is created in order to obtain several publications rather than just one. We are not necessarily accusing your team of this, but the way things have unfolded, does look similar. It may have been a misunderstanding in terms of how JOSS works, and what is acceptable for JOSS. It may be that to date you have not been used to proper git/github based forking and software development.
However, at this point it is important that we point out that this fragmented form of publishing the tensegrity related works is not the proper approach, and not acceptable for JOSS.

Given the overlap, the closely related functionality, and the shared codes, of this work OriTsgEFA and TsgDMC (as well as the former publications), we have decided to reject both current JOSS submissions (OriTsgEFA and TsgDMC). Instead, what we recommend, is that you merge all projects into a single tensegrity project. Ideally this should be done using the git/github methods available, such that the merged project will feature clear authorship from all contributors of the sub-projects. Once this merged project has been established, it should contain for instance a single `tenseg_plot.m` file shared by several functional codes/examples, it should then have a single clear README, a single contributing guideline document, and a single merged documentation set, and a single set of testing/example codes, etc. (some static, some dynamic, some featuring FEM, some data-based etc.).

Although the two current submission will be rejected, we would be open to you submitting the described single merged project for future consideration in JOSS. For such a re-submission we would ask that you clearly clarify the key differences/advances over the previous work. We would recommend that, prior to submission to JOSS, that you post this information in a pre-submission enquiry (e.g. here https://github.com/openjournals/joss/issues) where you address myself, so I can look if your project appears ready to be considered for JOSS.

I hope the above is clear. Let me know if you have questions.
