### 2023-10-19 #5939 @Muhao-Chen
@Kevin-Mattheus-Moerman Thank you for your prompt response and valuable feedback. We would like to address the issues you raised.

It is worth noting that none of my co-authors have previously authored any papers related to Tensegrity in collaboration with the authors listed in STEDY.

Regarding the overlap in functions, the primary areas affected are plotting and the material database, encompassing materials such as steel, aluminum, and wood. We are hesitant to undertake a complete rewrite of this section merely for the sake of consistency with prior plots and material selections. While we know the option of creating new plotting and material functions to avoid this overlap, we believe it may not be necessary. Instead, we prefer to focus our efforts on refining the core functions.

We intend to develop an independent package to ensure user convenience, such that new users do not need to download multiple packages and potentially get lost. Your suggestions in this regard are certainly appreciated. We will fork the existing project to accommodate your recommendations.

We don't have a big project, to be honest. It is essential to clarify that we are not trying to publish things piece by piece. Our primary focus is on Tensegrity research. Wherever our research goes, we open-source the most powerful and impactful tools resulting from our research works.

We take pride in OriTsgEFA, as it stands as the first package enabling the study of integrated Tensegrity and origami structures. Similarly, TsgDMC represents a groundbreaking tool for data-based modeling and controlling Tensegrity structures.

The rejection by JOSS has indeed been a discouragement for us, given our strong commitment to open-source contributions. Your suggestion of consolidating our projects into a single merged project is excellent. This implies the need for a well-defined long-term plan encompassing the theoretical aspects we intend to explore and the functions we aim to develop. We greatly appreciate your advice and assure you we will do our best.
